overhauls form state handling, removing unnecessary rerenders and deep comparisons
